adding a tandemPooObject
Will create new branch ase functionality and members of the the
PooObject are to change.
PooObject will no longer need an input component as tandem will be the
acting player.
This design change is to help decouple the interaction of player from
PooObject to tandemPooObj as pooObject doesnt need to know about player
input it is only resposible when in tandem form.